requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0
python-telegram-bot==20.6
